+++ CHANGELOG.md
@@ -5,6 +5,20 @@
 
 * integrate settings API for theme and language synchronization ([5189d72](https://github.com/carteracredit/workflow/commit/5189d72875cb8ccab2662d9ce0a1c75d72dc4214))
 
+# [1.2.0-rc.2](https://github.com/carteracredit/workflow/compare/v1.2.0-rc.1...v1.2.0-rc.2) (2026-04-18)
+
+
+### Features
+
+* enhance deployment environment handling in publish modal ([cb4ec55](https://github.com/carteracredit/workflow/commit/cb4ec5550a2fc52d1a5cbf1bfe2418bf451f1d1a))
+
+# [1.2.0-rc.1](https://github.com/carteracredit/workflow/compare/v1.1.0...v1.2.0-rc.1) (2026-04-17)
+
+
+### Features
+
+* integrate settings API for theme and language synchronization ([5189d72](https://github.com/carteracredit/workflow/commit/5189d72875cb8ccab2662d9ce0a1c75d72dc4214))
+
 # [1.1.0](https://github.com/carteracredit/workflow/compare/v1.0.0...v1.1.0) (2026-04-16)